Rename files, fix weightmap class
