+++ src/appleseed/CMakeLists.txt
@@ -1441,8 +1441,8 @@ set (renderer_modeling_aov_sources
     renderer/modeling/aov/aovfactoryregistrar.cpp
     renderer/modeling/aov/aovfactoryregistrar.h
     renderer/modeling/aov/aovtraits.h
-    renderer/modeling/aov/cryptomatte.cpp
-    renderer/modeling/aov/cryptomatte.h
+    renderer/modeling/aov/cryptomatteaov.cpp
+    renderer/modeling/aov/cryptomatteaov.h
     renderer/modeling/aov/denoiseraov.cpp
     renderer/modeling/aov/denoiseraov.h
     renderer/modeling/aov/depthaov.cpp
