+++ src/appleseed/renderer/modeling/aov/cryptomatteaov.h
